Refactor: renomeando arquivos do modal de confirmação de arquivar usuário
